==> analyses/02_data-management.Rmd <==
---
title: "Source Data Reproduction Attempt to Evaluate 6 Claims from Kim_SocSciMed_2016_AqDO"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
  code_folding: show
self_contained: true
highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")
library(pacman)
p_load(tidyverse, magrittr, scales, haven, lubridate, sjmisc, report)
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)
knitr::opts_knit$set(width = 125)
```

<!-- ------------------------------------------------------------ --> 

**Reproduction analyst:** Radoslaw Panczak  
**SCORE RR ID:** 65z92  
**OSF Project:** https://osf.io/7msd8  

# Data 

## Source 

The fourth wave of the Indonesia Family Life Survey (IFLS4) available for download [here](https://www.rand.org/well-being/social-and-behavioral-policy/data/FLS/IFLS/ifls4.html).  

Documentation is available [here](https://www.rand.org/well-being/social-and-behavioral-policy/data/FLS/IFLS/download.html).  

## Availability 

The data is available for download upon registration but the data custodians ask for not distributing it further so *the OSF [project component](https://osf.io/3xpnu/) that stores the copy of data (both raw and prepared) used in these analyses* has been kept ***private*** and the Github [repository](https://github.com/RPanczak/SCORE_Kim_SocSciMed_2016_AqDO) ***does not contain data either***.  

# Variables

## DOB, age, sex, blood pressure

From `bus1_1.dta` file:   

`us00`      Respondent measured?
*1. Yes*

`us01`      Sex  
1. Male
3. Female

`us03`      Age  

`us02yr`    Year of birth  

`us07bx`    Blood pressure (2nd measurement)  
`us07b1`    Blood pressure (2nd measurement)  
`us07b2`    Blood pressure (2nd measurement)  
`us07bp`    Blood pressure (2nd measurement)  

From `bus1_2.dta` file:    

`us07cx`    Blood pressure (3rd measurement)  
`us07c1`    Blood pressure (3rd measurement)  
`us07c2`    Blood pressure (3rd measurement)  
`us07cp`    Blood pressure (3rd measurement)  

Selecting:  

- only 'measured' respondents 
- respondents with information on sex and age
- respondents 45 years of age and above
- respondents with information on two blood pressure measurements (2nd & 3rd measurement).   

The second and third blood pressure measures were averaged and finally only hypertensive individuals were included  

```{r}
bus1_1 <- read_dta("data-raw/IFLS4/hh07_all_dta/bus1_1.dta") %>% 
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%
  select(hhid07_9, hhid07, pid07,
         us00,
         us01, us03,
         us02yr,
         us07bx, us07b1, us07b2) %>% 
  mutate_all(as.numeric) %>% 
  
  # only measured respondents
  filter(us00 == 1) %>% 
  select(-us00) %>% 
  
  # missing sex
  filter(!is.na(us01)) %>% 
  # recoding sex for regression
  mutate(sex = ifelse(us01 == 1, 0, 1)) %>% 
  select(-us01) %>% 
  
  # implausible DOB
  mutate(us02yr = if_else(us02yr >= 9998, NA_real_, us02yr)) %>% 
  # missing age
  filter(us03 < 998) %>% 
  # age inclusion
  filter(us03 >= 45) %>% 
  
  # bp1 measured
  filter(us07bx == 1) %>% 
  select(-us07bx) %>% 
  # missing bp
  filter(us07b1 != 999) %>% 
  filter(us07b2 != 999) 

bus1_2 <- read_dta("data-raw/IFLS4/hh07_all_dta/bus1_2.dta") %>% 
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%
  select(hhid07, pid07,
         us07cx, us07c1, us07c2) %>% 
  mutate_all(as.numeric) %>% 
  
  # bp2 measured
  filter(us07cx == 1) %>% 
  
  # missing bp
  filter(us07c1 != 999) %>% 
  filter(us07c2 != 999) %>% 
  select(-us07cx)

bus_1 <- 
  inner_join(bus1_1, bus1_2) %>%
  
  # average of two bp measures (2nd and 3rd)
  rowwise() %>% 
  mutate(us071 = mean(c(us07b1, us07c1), na.rm = FALSE),
         us072 = mean(c(us07b2, us07c2), na.rm = FALSE)) %>% 
  ungroup() %>% 
  select(-us07b1, -us07c1, -us07b2, -us07c2) %>% 
  
  # selecting hypertensive
  filter(us071 > 140 | us072 > 90) %>% 
  select(-us071, -us072) %>%
  
  # misc
  rename(age = us03) %>% 
  mutate(age2 = age*age) %>% 
  relocate(age2, .after = age)
```

```{r include=FALSE}
rm(bus1_1, bus1_2)
```

```{r eval=FALSE, include=FALSE}
summary(bus_1$age)
```

After this steps end up with `r number(nrow(bus_1), big.mark = ",")` individuals in the dataset which is already diverging from the analysis dataset reported in the paper (N = 4209).  

The sex distribution seems to follow what is reported in the paper:  

```{r echo=FALSE}
frq(bus_1, sex)
```

## Hypertension diagnosed

From `b3b_cd3.dta` file:  

`cdtype`    Type chronic conditions  
A. Hypertension  

`cd05`    Has a doctor ever told you that you had  
1. Yes   

```{r}
b3b_cd3  <- read_dta("data-raw/IFLS4/hh07_all_dta/b3b_cd3.dta") %>% 
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%
  select(pid07, hhid07,
         cd05, cdtype) %>% 
  filter(cdtype == "A") %>% 
  select(-cdtype) %>% 
  mutate_all(as.numeric) %>% 
  # excluding missing
  filter(cd05 != 9) %>% 
  # outcome is NO diagnosis
  mutate(cd05 = ifelse(cd05 == 1, 0, cd05)) %>% 
  mutate(cd05 = ifelse(cd05 == 3, 1, cd05)) %>% 
  rename(underdiagnosed = cd05)
```

Proportion of under-diagnosed is higher but this is still dataset without link to the selection of patients made above:  

```{r echo=FALSE}
frq(b3b_cd3, underdiagnosed)
```

After the link the proportion is very similar to the one reported in paper but we have a significant amount of missing:  

```{r echo=FALSE}
bus_1 %>% 
  left_join(b3b_cd3) %>% 
  frq(underdiagnosed)
```

Selecting only under-diagnosed patients counts of sexes resemble these presented in paper:  

```{r echo=FALSE}
bus_1 %>% 
  left_join(b3b_cd3) %>% 
  filter(underdiagnosed == 1) %>% 
  frq(sex)
```

## Household expenditures

From `b1_ks2.dta` file.   

`ks06`    nonfood expenditure last month (Rp)  

Multiple types of expenditures are aggregated by household.  

```{r}
b1_ks2  <- read_dta("data-raw/IFLS4/hh07_all_dta/b1_ks2.dta") %>% 
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%
  select(hhid07,
         ks06) %>% 
  mutate_all(as.numeric) %>% 
  group_by(hhid07) %>% 
  summarize(expenditures = sum(ks06)) %>% 
  ungroup()
```

```{r eval=FALSE, include=FALSE}
summary(b1_ks2$expenditures)
```

We have few households where expenditures are missing. And also - where they are equal zero - the latter will create problems where log of this variable needed for regression is calculated:  

```{r echo=FALSE}
bus_1 %>% 
  left_join(b1_ks2) %>% 
  frq(expenditures == 0)
```

## Health status

From `b3b_kk1.dta` file:  

`kk01`    Generally how is your health?  

1. Very healthy
2. Fairly healthy
3. In poor health
4. Very sick

```{r}
b3b_kk1  <- read_dta("data-raw/IFLS4/hh07_all_dta/b3b_kk1.dta") %>% 
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%
  select(pid07, hhid07,
         kk01) %>% 
  mutate_all(as.numeric) %>% 
  mutate(poor_health = ifelse(kk01 >= 3, 1, 0)) %>% 
  select(-kk01)
```

```{r eval=FALSE, include=FALSE}
frq(b3b_kk1, poor_health)
```

There is a poor link between this dataset and the `bus_1` file which results in large amount of missings:     

```{r echo=FALSE}
bus_1 %>% 
  left_join(b3b_kk1) %>% 
  frq(poor_health)
```

## Distance to health facilities

Again paper is rather cryptic and specifies it only as:  

> the distance from the closest health center (to proxy for the ease of access to medical care)  

The closest variables that would match this description are found in the `b1_pp.dta` dataset:  

`pp5`       One way travel time (unit of measurement)  

`pptype`    Type of out-patient facility  
C. Public Health Center/Auxiliary Center   

(Note:  it's important to take into account `pp5x` unit of measurement to make them comparable).  

```{r}
b1_pp  <- read_dta("data-raw/IFLS4/hh07_all_dta/b1_pp.dta") %>% 
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%
  select(hhid07_9,
         pptype,
         pp5, pp5x) %>% 
  # facility type
  filter(pptype == "C") %>% 
  select(-pptype) %>% 
  mutate_all(as.numeric) %>% 
  # exclude missings
  filter(pp5x < 30) %>% 
  # standardize other units to minutes
  mutate(pp5 = ifelse(pp5x == 2, pp5*60, pp5)) %>% 
  mutate(pp5 = ifelse(pp5x == 3, pp5*24*60, pp5)) %>% 
  select(-pp5x) %>% 
  rename(travel_time = pp5)
```

```{r eval==FALSE}
summary(b1_pp$travel_time)
```

Once again, we have some missing information here:  

```{r echo=FALSE}
bus_1 %>% 
  left_join(b1_pp) %>% 
  frq(is.na(travel_time))
```

## Education

Paper defines it laconically as: 

> measured in years of formal education

From `b3a_dl1.dta` file we can grab:    

`dl04`        Have you ever attended school?

`dl05a`       At what age did you enter school  

`dl06`        Highest level of education attended

`dl07byr`     Year graduated or left school  

`dl07d`       IVWR NOTE: Still in school
1. Yes

These together with `dob` these variables can be used to obtain crude years of education.  

```{r}
b3a_dl1 <- read_dta("data-raw/IFLS4/hh07_all_dta/b3a_dl1.dta") %>% 
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%
  select(pid07, hhid07,
         dl04,
         dl06,
         dl07d, 
         dl05a, dl07byr) %>% 
  mutate_all(as.numeric) %>% 
  # if still in school replace to 2007, year of survey
  mutate(dl07byr = ifelse (is.na(dl07byr) & dl07d == 1, 2007, dl07byr)) %>% 
  # missings
  filter(dl07byr < 9998 | is.na(dl07byr)) %>%
  filter(dl05a < 98 | is.na(dl07byr)) %>% 
  select(-dl07d)
```

Of course - this is only possible for patients that have these variables non missing which seems to be quite a large problem:  

```{r eval=FALSE, include=FALSE}
bus_1 %>% 
  left_join(b3a_dl1) %>% 
  frq(is.na(dl07byr))
```

## Risk & time preference

> For the time and risk preference parameters, we follow Ng (2013) and group respondents in four distinct groups from the most patient to the most impatient, respectively from the least risk averse to the most risk averse.   

Paper of Ng defines two measures of risk and two measures of time and does not specify which of them were used and in which form (one vs. the other? combined?).  

Here we prepare all 4 measures (2 for risk and two for time).  

Method of Ng basically takes into account answers to five questions and follows a flowchart to assign the individual to a risk category ranging from 0 to 4.  

From the dataset `b3a_si.dta` we can use variables:  

`si01`               Which money earning option choosen?
`si02`               Are you sure?
`si03`               Which option will you choose?
`si04`               Which option will you choose?
`si05`               Which option will you choose?

to construct risk measure 1.  

```{r}
risk1 <- read_dta("data-raw/IFLS4/hh07_all_dta/b3a_si.dta") %>%
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%  
  select(-version, -module, -pidlink) %>% 
  mutate_all(as.numeric) %>% 
  relocate(hhid07, .after = pid07) %>% 
  filter(si01 < 8) %>% 
  filter(si02 < 8 | is.na(si02)) %>% 
  filter(si03 < 8 | is.na(si03)) %>% 
  filter(si04 < 8 | is.na(si04)) %>% 
  filter(si05 < 8 | is.na(si05)) %>% 
  mutate(
    risk1 = case_when(
      si01 == 1 & si02 == 1    ~ 0,
      si03 == 2 & si05 == 2    ~ 1,
      si03 == 2 & si05 == 1    ~ 2,
      si03 == 1 & si04 == 2    ~ 3,
      si03 == 1 & si04 == 1    ~ 4
    )
  ) %>% 
  select(pid07, hhid07, risk1)
```

Then in a similar vein variables `si11` to `si15` can be used to construct risk measure 2.  

```{r}
risk2 <- read_dta("data-raw/IFLS4/hh07_all_dta/b3a_si.dta") %>%
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%  
  select(-version, -module, -pidlink) %>% 
  mutate_all(as.numeric) %>% 
  relocate(hhid07, .after = pid07) %>% 
  filter(si11 < 8) %>% 
  filter(si12 < 8 | is.na(si12)) %>% 
  filter(si13 < 8 | is.na(si13)) %>% 
  filter(si14 < 8 | is.na(si14)) %>% 
  filter(si15 < 8 | is.na(si15)) %>% 
  mutate(
    risk2 = case_when(
      si11 == 1 & si12 == 1    ~ 0,
      si13 == 2 & si15 == 2    ~ 1,
      si13 == 2 & si15 == 1    ~ 2,
      si13 == 1 & si14 == 2    ~ 3,
      si13 == 1 & si14 == 1    ~ 4
    )
  ) %>% 
  select(pid07, hhid07, risk2)
```

```{r eval=FALSE, include=FALSE}
frq(risk1, risk1)
frq(risk2, risk2)
```

```{r echo=FALSE}
bus_1 %>% 
  left_join(risk1) %>% 
  left_join(risk2) %>% 
  frq(risk1, risk2)
```

Combining two risk scores into one we can impute missing values of first measure with non-missing ones from two:  

```{r echo=FALSE}
bus_1 %>% 
  left_join(risk1) %>% 
  left_join(risk2) %>% 
  frq(is.na(risk1) & !is.na(risk2))
```

In a very similar procedure variables `si21a` to `si21d` and `si22a` and `si22d` can be used to derive time preference 1 and 2 respectively.   

```{r}
time1 <- read_dta("data-raw/IFLS4/hh07_all_dta/b3a_si.dta") %>%
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%  
  select(-version, -module, -pidlink) %>% 
  mutate_all(as.numeric) %>% 
  relocate(hhid07, .after = pid07) %>% 
  filter(si21a < 8) %>% 
  filter(si21b < 8 | is.na(si21b)) %>% 
  filter(si21c < 8 | is.na(si21c)) %>% 
  filter(si21d < 8 | is.na(si21d)) %>% 
  filter(si21e < 8 | is.na(si21e)) %>% 
  mutate(
    time1 = case_when(
      si21a == 2                 ~ 0,
      si21b == 2 & si21d == 2    ~ 1,
      si21b == 2 & si21d == 1    ~ 2,
      si21b == 1 & si21c == 2    ~ 3,
      si21b == 1 & si21c == 1    ~ 4
    )
  ) %>% 
  select(pid07, hhid07, time1)
```

Then again in a similar vein variables `si22a` to `si22e` can be used to construct risk measure 2.  

```{r}
time2 <- read_dta("data-raw/IFLS4/hh07_all_dta/b3a_si.dta") %>%
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%  
  select(-version, -module, -pidlink) %>% 
  mutate_all(as.numeric) %>% 
  relocate(hhid07, .after = pid07) %>% 
  filter(si22a < 8) %>% 
  filter(si22b < 8 | is.na(si22b)) %>% 
  filter(si22c < 8 | is.na(si22c)) %>% 
  filter(si22d < 8 | is.na(si22d)) %>% 
  filter(si22e < 8 | is.na(si22e)) %>% 
  mutate(
    time2 = case_when(
      si22a == 2                 ~ 0,
      si22b == 2 & si22d == 1    ~ 1,
      si22b == 2 & si22d == 2    ~ 2,
      si22b == 1 & si22c == 2    ~ 3,
      si22b == 1 & si22c == 1    ~ 4
    )
  ) %>% 
  select(pid07, hhid07, time2)
```

```{r eval=FALSE, include=FALSE}
frq(time1, time1)
frq(time2, time2)
```

```{r echo=FALSE}
bus_1 %>% 
  left_join(time1) %>% 
  left_join(time2) %>% 
  frq(time1, time2)
```

```{r eval=FALSE, include=FALSE}
# nothing to rescue 
bus_1 %>% 
  left_join(time1) %>% 
  left_join(time2) %>% 
  frq(is.na(time1) & !is.na(time2))
```

## Disease management

Information about this variable from the paper:  

> We construct a dummy variable that is equal to 1 if respondents checked their blood pressure at least once in the month preceding the interview.  

> Disease management does not apply to individuals who are not aware of having any disease, so we restrict our sample to those respondents who were previously diagnosed with hypertension.  

From `b3b_rj3.dta` file we can grab:    

`rj24amth`    Last blood pressure check (month)  
`rj24ayr`     Last blood pressure check (year)  


```{r}
b3b_rj3 <- read_dta("data-raw/IFLS4/hh07_all_dta/b3b_rj3.dta") %>% 
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%
  select(pid07, hhid07_9,
         rj24amth, rj24ayr) %>% 
  mutate_all(as.numeric) %>% 
  # missings
  filter(!is.na(rj24amth)) %>% 
  filter(!is.na(rj24ayr)) %>% 
  filter(rj24amth < 98) %>% 
  filter(rj24ayr < 9998) %>% 
  mutate(bp_date = ymd(paste(rj24ayr, rj24amth, 1, sep= " "))) 
```

From the file `b1_cov.dta` we can grab:  

`ivwday1`       Date of intrvw (day of month)(1)
`ivwmth1`       Date of interview (month)(1)
`ivwyr1`        Date of Visit (year)(1)

```{r}
b1_cov <- read_dta("data-raw/IFLS4/hh07_all_dta/b1_cov.dta") %>% 
  # zap_label() %>% 
  zap_labels() %>% zap_formats() %>%
  select(pid07, hhid07_9,
         # ivwday1, # fixing 1st to correctly get differences
         ivwmth1, ivwyr1) %>% 
  mutate_all(as.numeric) %>% 
  filter(!is.na(ivwmth1)) %>% 
  filter(!is.na(ivwyr1)) %>% 
  mutate(ivwyr1 = ivwyr1 + 2000) %>% 
  mutate(iv_date = ymd(paste(ivwyr1, ivwmth1, 1, sep= " "))) 
```

These dates can be then compared to generate an indicator of measurement in the month proceeding the interview.  

Note: values of BP measurement that are dated **after** date of interview are considered implausible and do not count towards the positive indicator.  

```{r}
b3b_rj3 %<>% inner_join(b1_cov) %>% 
  mutate(bp_dist = as.numeric(iv_date - bp_date)) %>% 
  mutate(bp_month = if_else(bp_dist > 0 & bp_dist <= 31, 1, 0)) %>% 
  select(-ivwyr1, -ivwmth1,
         - rj24ayr, -rj24amth,
         -iv_date, -bp_date,
         -bp_dist)

rm(b1_cov)
```

```{r eval=FALSE, include=FALSE}
frq(b3b_rj3, bp_month)
```

There is a very big amount of information missing here:  

```{r echo=FALSE}
bus_1 %>% 
  left_join(b3b_rj3) %>% 
  frq(bp_month)
```

# Final dataset

## Merging together

Due to very large amount of missing information on the number of years of education, categorical variable of highest education achieved was used to impute values into simple four categories of 'elementary', 'junior high', 'senior high' and 'university'.  

Similarly, large amount of missings in the disease management indicator was assumed to mean 'no'.  

Last but not least, risk and time preferences are combined across two variables:  
- if risk 1 is missing and 2 no then 2 is used
- if risk 1 is zero then two is used instead

Zero and missing values of such constructed variables are then excluded.  

```{r}
data <- bus_1 %>% 
  
  # hyper diagnosis
  left_join(b3b_cd3) %>% 
  # exclude missings
  filter(!is.na(underdiagnosed)) %>% 
  
  # expenditures
  left_join(b1_ks2) %>% 
  # exclude missings
  filter(!is.na(expenditures)) %>% 
  # avoid 0s for log
  mutate(expenditures = if_else(expenditures == 0, 1, expenditures)) %>% 
  # log
  mutate(expenditures_log = log(expenditures)) %>% 
  select(-expenditures) %>% 
  
  # poor health
  left_join(b3b_kk1) %>% 
  # exclude missings
  filter(!is.na(poor_health)) %>%  
  
  # distance
  left_join(b1_pp) %>% 
  # exclude missings
  filter(!is.na(travel_time)) %>% 
  
  # education
  left_join(b3a_dl1) %>% 
  # calculate years of education
  mutate(education = dl07byr - us02yr - dl05a) %>% 
  select(-dl07byr, -us02yr, -dl05a) %>% 
  # zero for those who never attended
  mutate(education = if_else(dl04 == 3, 0, education)) %>% 
  select(-dl04) %>% 
  # imputing education from cateogricals
  # elementary school to 6
  mutate(education = if_else(is.na(education) & dl06 == 2, 6, education)) %>% 
  # junior high to 9
  mutate(education = if_else(is.na(education) & 
                               (dl06 == 3 | dl06 == 4), 9, education)) %>% 
  # senior high to 12
  mutate(education = if_else(is.na(education) & 
                               (dl06 == 5 | dl06 == 6), 12, education)) %>% 
  # uni to 16
  mutate(education = if_else(is.na(education) & 
                               (dl06 == 13 | dl06 == 60 | dl06 == 61), 16, education)) %>%   select(-dl06) %>% 
  filter(!is.na(education)) %>%
  filter(education >= 0) %>%
  
  # time
  left_join(time1) %>% 
  left_join(time2) %>% 
  # imputing missings in 1 from 2
  mutate(time = ifelse(!is.na(time1), time1, time2)) %>% 
  # imputing 0 category in 1 from 2
  mutate(time = ifelse(time1 == 0, time2, time1)) %>% 
  filter(!is.na(time)) %>% 
  filter(time > 0) %>% 
  relocate(time, .before = time1) %>% 
  select(-time1, -time2) %>% 
  
  # risk
  left_join(risk1) %>% 
  left_join(risk2) %>% 
  # imputing missings in 1 from 2
  mutate(risk = ifelse(!is.na(risk1), risk1, risk2)) %>% 
  # imputing 0 category in 1 from 2
  mutate(risk = ifelse(risk1 == 0, risk2, risk1)) %>% 
  filter(!is.na(risk)) %>% 
  filter(risk > 0) %>% 
  relocate(risk, .before = risk1) %>% 
  select(-risk1, -risk2) %>% 
  
  # disease management
  left_join(b3b_rj3) %>% 
  # second version imputing missing to zero (ie. 'no')
  mutate(bp_month2 = ifelse(is.na(bp_month), 0, bp_month)) %>% 
  select(-bp_month)
```

```{r include=FALSE}
rm(bus_1, b3b_cd3, b1_ks2, b3b_kk1, b1_pp, b3a_dl1, b3b_rj3,
   risk1, risk2, time1, time2); gc()
```

## Dataset

```{r echo=FALSE}
data %>% 
  select(-hhid07_9, -hhid07, - pid07) %>% 
  mutate(sex = factor(sex),
         underdiagnosed = factor(underdiagnosed),
         poor_health = factor(poor_health),
         bp_month2 = factor(bp_month2)) %>% 
  report::report()
```

## Codebook

- `hhid07_9`          household ID  
- `hhid07`            household ID  
- `pid07`             person ID  
- `age`               age of individual  
- `age2`              age squared  
- `sex`               sex of individual (0 = female, 1 = male)  
- `underdiagnosed`    binary indicator of being under-diagnosed with blood pressure (0 = no, 1 = yes)
- `expenditures_log`  nonfood expenditures of household during last month (logged)  
- `poor_health`       binary indicator of being in poor health (0 = no, 1 = yes)   
- `travel_time`       distance from the closest health centre  
- `education`         years of education (imputed)  
- `time`              time preference (1-4 scale, see Ng 2012 for details)  
- `risk`              risk preference (1-4 scale, see Ng 2012 for details)  
- `bp_month2`         blood pressure measurement during last month (0 = no, 1 = yes; imputed)

## Summaries

Various tabulations and summaries of variables used for selecting patients and running regression models.  

### Sex, poor health & under-diagnosed

```{r}
frq(data, sex)

frq(data, poor_health)

frq(data, underdiagnosed)

gmodels::CrossTable(data$underdiagnosed, data$poor_health,
                    digits = 2, prop.chisq = FALSE)
```

### Continuous variables 

```{r}
summary(data$age)

summary(data$education)

summary(data$expenditures_log)

summary(data$travel_time)
```

### Time & risk 

```{r}
frq(data, risk)

# gmodels::CrossTable(data$risk1, data$risk2,
#                     digits = 2, prop.chisq = FALSE)

frq(data, time)

# gmodels::CrossTable(data$time1, data$time2,
#                     digits = 2, prop.chisq = FALSE)
```

### Under-diagnosed subsample 

```{r}
data_diag <- data %>% 
  filter(underdiagnosed == 0)

# frq(data_diag, bp_month)
frq(data_diag, bp_month2)

# gmodels::CrossTable(data_diag$sex, data_diag$bp_month,
#                     digits = 2, prop.chisq = FALSE)

gmodels::CrossTable(data_diag$sex, data_diag$bp_month2,
                    digits = 2, prop.chisq = FALSE)
```

```{r include=FALSE}
write_rds(data, "data/data.Rds")
```

<!-- ----------------------------------------------------- -->

# References

Kim Y, Radoias V (2016) Education, individual time preferences, and asymptomatic disease detection. *Social Science & Medicine* **150**, 15-22, https://doi.org/10.1016/j.socscimed.2015.11.051.  

Ng J (2012) Elicited Risk and Time Preferences: The Role of Demographics, Cognition, and Interviewers. Working paper. University of Southern California. http://static1.1.sqspcdn.com/static/f/884336/20238159/1347396545633/ng_dev_seminar9-10.pdf  

<!-- ----------------------------------------------------- -->

# Computing Environment

```{r echo=FALSE, results='asis'}
report(sessionInfo())
```
