==> analyses/03_data-analysis.Rmd <==
---
title: "Source Data Reproduction Attempt to Evaluate 6 Claims from Kim_SocSciMed_2016_AqDO"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
  code_folding: show
self_contained: true
highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")
library(pacman)
p_load(tidyverse, scales, magrittr, margins, report, googlesheets4)
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)
knitr::opts_knit$set(width = 125)
```

<!-- ------------------------------------------------------------ --> 

**Reproduction analyst:** Radoslaw Panczak  
**SCORE RR ID:** 65z92  
**OSF Project:** https://osf.io/7msd8  

# Transparency Trail 

## Preregistration

Details stored in publicly available Goole Doc:  

```{r}
sheet_location = "https://docs.google.com/spreadsheets/d/1lqwu5mGKe4URtAgyd0uAOJAU3YmsyKRaLx6eWYicUdY/edit#gid=226716288"
```

**Note:** *the registration details in tables below have been entered post-analysis.*  

## Link to analysis script(s)

**OSF:** https://osf.io/z4nq5/  
**GitHub:** https://github.com/RPanczak/SCORE_Kim_SocSciMed_2016_AqDO/blob/main/analyses/03_data-analysis.Rmd  

## Data 

```{r include=FALSE}
data <- read_rds("data/data.Rds")
```

The steps used to construct this dataset from the original data are are reproducible using `02_data-management.Rmd` script. The output of these steps together with comments and detail is documented in the `02_data-management.html` file which can be obtained from the mentioned `Rmd` file.  

Prepared dataset `data.Rds` is stored in `/data` directory (both in OSF project and on GitHub) and consists of `r number(nrow(data), big.mark = ",")` observations. This is only `r percent(nrow(data)/4209, accuracy = 0.1)` of the size of the full sample reported in the paper.  

*Sample size alone already violates any reproducibility criteria and would most likely classify this study as not possible to be reproduced.*  

There were several steps that were poorly or completely not reported in the original study that would be necessary to reconstruct the dataset. For instance Authors are silent on their treatment of missing data that are present in many variables. In case of 'simple' variables such as age and sex I defaulted to the simplest solution of complete case analysis since there is no information on how else Authors would approach that. On a similar not Authors do not report what is their approach to handle implausible values (year of birth) or treatment of zero values of expenditures when logarithm is taken.  

The problem of missing information is particularly important when it comes to education (which is one of the key focus areas of the paper!).  Information is missing in order to construct variable indicating years of education without even slightest hint on what has been done to mitigate that. The only reasonable solution that I was able to come up with was to impute values into simple four categories of 'elementary', 'junior high', 'senior high' and 'university'. Without that step hardly any analyses would be possible.  

Similarly, there were large amount of missings in the disease management indicator (also important for the analyses) - the missing information here was assumed to mean 'no'. 

Importanly, methods of construction of risk and time preferences indicators was not described. Authors refer the reader to paper of Ng (2013) that defines two measures of risk and two measures of time and does not specify which of them should be used and in which form (one vs. the other? combined?). In order to mitigate problem of missing information I used all four measures (two for risk and two for time) and combine them to be able to include as many study participants as possible.  

Last but not least Authors do not give much detail on their modelling strategy. Results in Table 2 & 4 are presented for "Aggregate sample" and then two strata. It is not clear if the coefficients of stratified analyses were obtained from running separate models on two selected subpopulations or another method. I assumed the first case.  

## General modelling details

Probit models are fitted with `glm` function from the `stats` package (part of base R).  

Marginal effects were obtained using `margins` package.  

Complete details on the software & computing environment are given in the dedicated section at the bottom of this document.  

## Analysis attempt 1

The first modelling followed general ideas regarding data preparation and modelling described above but *only selected respondents in poor health*.  

Completed claims: `single-trace`, `21v1on`, `dx1x5w`

The second modelling *only selected respondents in good health*.  

Completed claims: `ok8kxo`, `n3k3qz`  

The third modelling *only selected respondents in good health with diagnosed hypertension*.   

Completed claim: `8rlrky`  

<!-- ------------------------------------------------------------ --> 

# Claim evaluations

## Claim ID: single-trace

### Coded claim 4 text (original paper)

> Education matters for these people, as more educated persons generally have higher opportunity costs of feeling sick and hence value their health higher, which pushes them harder to look for a cure in a doctor’s office (marginal effect for Years of Education = -0.00867, SE = 0.00420, significant at 5% level).  

### Reproduction data source(s)

https://www.rand.org/well-being/social-and-behavioral-policy/data/FLS/IFLS/ifls4.html  

### Description of reproduction data 

The fourth wave of the Indonesia Family Life Survey (IFLS4) is a continuation of IFLS, expanding the panel to 2007/2008. Downloaded from the link in the cell above on 2022-01-06; uploaded to OSF to ‘Data/data-raw’ directory.  

Selecting respondents in poor health only.  

```{r}
data_poor_health <- data %>% 
  filter(poor_health == 1) %>% 
  select(-poor_health)
```

Dataset consists of `r number(nrow(data_poor_health), big.mark = ",")` observations. 

### Model 

```{r}
claim_ok8kxo <- glm(underdiagnosed ~ 
                      age + age2 + education +
                      expenditures_log + time + risk +
                      travel_time + sex, 
                    family = binomial(link = "probit"), 
                    data = data_poor_health)

# summary(claim_ok8kxo)
summary(margins(claim_ok8kxo))
```

### Primary reproduction criteria

```{r}
primary <- read_sheet(sheet_location,
                      range = "single_trace!1:6") %>% 
  slice(-1) %>% 
  mutate(`Reproduction result` = "")

primary$`Reproduction result`[1] <- number(nrow(data_poor_health))

primary$`Reproduction result`[3] <- summary(margins(claim_ok8kxo)) %>% 
  filter(factor == "education") %>% 
  select(AME) %>% 
  mutate(AME = number(AME, accuracy = 0.000001)) %>% 
  pluck(1)

primary %>% 
  kableExtra::kable()
```

### Secondary reproduction criteria

```{r}
secondary <- read_sheet(sheet_location,
                        range = "single_trace!1:10") %>% 
  slice(8:9) %>% 
  mutate(`Reproduction result` = "")

secondary$`Reproduction result`[1] <- summary(margins(claim_ok8kxo)) %>% 
  filter(factor == "education") %>% 
  select(SE) %>% 
  mutate(SE = number(SE, accuracy = 0.000001)) %>% 
  pluck(1)

secondary$`Reproduction result`[2] <- summary(margins(claim_ok8kxo)) %>% 
  filter(factor == "education") %>% 
  select(p) %>% 
  mutate(p = number(p, accuracy = 0.000001)) %>% 
  pluck(1)

secondary %>% 
  kableExtra::kable()
```

### Analyst success criteria

All outcomes meet the precise reproduction criteria

### Reproduction outcome

Based on the criteria above, the claim **did not** reproduce.  


## Claim ID: 21v1on

### Coded claim 4 text (original paper)

> Table 2 contains the probit regression results for the determinants of hypertension under-diagnosis… For individuals in good health, the education level does not matter at all… [Table 2, Respondents in good health, Years of Education: Marginal effect = -0.00295, SE = 0.00206].  
### Reproduction data source(s)

https://www.rand.org/well-being/social-and-behavioral-policy/data/FLS/IFLS/ifls4.html  

### Description of reproduction data 

The fourth wave of the Indonesia Family Life Survey (IFLS4) is a continuation of IFLS, expanding the panel to 2007/2008. Downloaded from the link in the cell above on 2022-01-06; uploaded to OSF to ‘Data/data-raw’ directory.  

Selecting respondents in good health only.  

```{r}
data_good_health <- data %>% 
  filter(poor_health == 0) %>% 
  select(-poor_health)
```

Dataset consists of `r number(nrow(data_good_health), big.mark = ",")` observations.  

### Model 

```{r}
claim_21v1on <- glm(underdiagnosed ~ 
                      age + age2 + education +
                      expenditures_log + time + risk +
                      travel_time + sex, 
                    family = binomial(link = "probit"), 
                    data = data_good_health)

# summary(claim_21v1on)
summary(margins(claim_21v1on))
```

### Primary reproduction criteria

```{r}
primary <- read_sheet(sheet_location,
                      range = "Claim_ID_21v1on!1:6") %>% 
  slice(-1) %>% 
  mutate(`Reproduction result` = "")

primary$`Reproduction result`[1] <- number(nrow(data_good_health))

primary$`Reproduction result`[3] <- summary(margins(claim_21v1on)) %>% 
  filter(factor == "education") %>% 
  select(AME) %>% 
  mutate(AME = number(AME, accuracy = 0.000001)) %>% 
  pluck(1)

primary %>% 
  kableExtra::kable()
```

### Secondary reproduction criteria

```{r}
secondary <- read_sheet(sheet_location,
                        range = "Claim_ID_21v1on!1:10") %>% 
  slice(8:9) %>% 
  mutate(`Reproduction result` = "")

secondary$`Reproduction result`[1] <- summary(margins(claim_21v1on)) %>% 
  filter(factor == "education") %>% 
  select(SE) %>% 
  mutate(SE = number(SE, accuracy = 0.000001)) %>% 
  pluck(1)

secondary$`Reproduction result`[2] <- summary(margins(claim_21v1on)) %>% 
  filter(factor == "education") %>% 
  select(p) %>% 
  mutate(p = number(p, accuracy = 0.000001)) %>% 
  pluck(1)

secondary %>% 
  kableExtra::kable()
```

### Analyst success criteria

All outcomes meet the precise reproduction criteria

### Reproduction outcome

Based on the criteria above, the claim **did not** reproduce.  


## Claim ID: ok8kxo

### Coded claim 4 text (original paper)

> Table 2 contains the probit regression results for the determinants of hypertension under-diagnosis…On the other hand, people who are in poor overall health will feel the need to go see a doctor… [Table 2, Respondents in poor health, Years of Education: Marginal effect = -0.00867, SE = 0.00420, significant at 5% level].   

### Reproduction data source(s)

https://www.rand.org/well-being/social-and-behavioral-policy/data/FLS/IFLS/ifls4.html  

### Description of reproduction data 

The fourth wave of the Indonesia Family Life Survey (IFLS4) is a continuation of IFLS, expanding the panel to 2007/2008. Downloaded from the link in the cell above on 2022-01-06; uploaded to OSF to ‘Data/data-raw’ directory.  

Selecting respondents in poor health only.  

Dataset consists of `r number(nrow(data_poor_health), big.mark = ",")` observations.  

### Model 

Please note that model `claim_ok8kxo` from the *single-trace* claim described above is used here again.  

```{r}
# summary(claim_ok8kxo)
summary(margins(claim_ok8kxo))
```

### Primary reproduction criteria

```{r}
primary <- read_sheet(sheet_location,
                      range = "Claim_ID_ok8kxo!1:6") %>% 
  slice(-1) %>% 
  mutate(`Reproduction result` = "")

primary$`Reproduction result`[1] <- number(nrow(data_poor_health))

primary$`Reproduction result`[3] <- summary(margins(claim_ok8kxo)) %>% 
  filter(factor == "education") %>% 
  select(AME) %>% 
  mutate(AME = number(AME, accuracy = 0.000001)) %>% 
  pluck(1)

primary %>% 
  kableExtra::kable()
```

### Secondary reproduction criteria

```{r}
secondary <- read_sheet(sheet_location,
                        range = "Claim_ID_ok8kxo!1:10") %>% 
  slice(8:9) %>% 
  mutate(`Reproduction result` = "")

secondary$`Reproduction result`[1] <- summary(margins(claim_ok8kxo)) %>% 
  filter(factor == "education") %>% 
  select(SE) %>% 
  mutate(SE = number(SE, accuracy = 0.000001)) %>% 
  pluck(1)

secondary$`Reproduction result`[2] <- summary(margins(claim_ok8kxo)) %>% 
  filter(factor == "education") %>% 
  select(p) %>% 
  mutate(p = number(p, accuracy = 0.000001)) %>% 
  pluck(1)

secondary %>% 
  kableExtra::kable()
```

### Analyst success criteria

All outcomes meet the precise reproduction criteria

### Reproduction outcome

Based on the criteria above, the claim **did not** reproduce.  


## Claim ID: n3k3qz

### Coded claim 4 text (original paper)

> Table 2 contains the probit regression results for the determinants of hypertension under-diagnosis… for respondents in poor health, the time preference parameter and distance to health center do not matter. [Table 2, Respondents in poor health, Time preference: Marginal effect = 0.03429, SE = 0.02379].  

### Reproduction data source(s)

https://www.rand.org/well-being/social-and-behavioral-policy/data/FLS/IFLS/ifls4.html  

### Description of reproduction data 

The fourth wave of the Indonesia Family Life Survey (IFLS4) is a continuation of IFLS, expanding the panel to 2007/2008. Downloaded from the link in the cell above on 2022-01-06; uploaded to OSF to ‘Data/data-raw’ directory.    

Selecting respondents in poor health only.  

Dataset consists of `r number(nrow(data_poor_health), big.mark = ",")` observations.  

### Model 

See model `claim_ok8kxo` above.    

```{r}
summary(margins(claim_ok8kxo))
```

### Primary reproduction criteria

```{r}
primary <- read_sheet(sheet_location,
                      range = "Claim_ID_n3k3qz!1:6") %>% 
  slice(-1) %>% 
  mutate(`Reproduction result` = "")

primary$`Reproduction result`[1] <- number(nrow(data_poor_health))

primary$`Reproduction result`[3] <- summary(margins(claim_ok8kxo)) %>% 
  filter(factor == "time") %>% 
  select(AME) %>% 
  mutate(AME = number(AME, accuracy = 0.000001)) %>% 
  pluck(1)

primary %>% 
  kableExtra::kable()
```

### Secondary reproduction criteria

```{r}
secondary <- read_sheet(sheet_location,
                        range = "Claim_ID_n3k3qz!1:10") %>% 
  slice(8:9) %>% 
  mutate(`Reproduction result` = "")

secondary$`Reproduction result`[1] <- summary(margins(claim_ok8kxo)) %>% 
  filter(factor == "time") %>% 
  select(SE) %>% 
  mutate(SE = number(SE, accuracy = 0.000001)) %>% 
  pluck(1)

secondary$`Reproduction result`[2] <- summary(margins(claim_ok8kxo)) %>% 
  filter(factor == "time") %>% 
  select(p) %>% 
  mutate(p = number(p, accuracy = 0.000001)) %>% 
  pluck(1)

secondary %>% 
  kableExtra::kable()
```

### Analyst success criteria

All outcomes meet the precise reproduction criteria

### Reproduction outcome

Based on the criteria above, the claim **did not** reproduce.  


## Claim ID: dx1x5w

### Coded claim 4 text (original paper) 

> Table 2 contains the probit regression results for the determinants of hypertension under-diagnosis…For people in good overall health, the decision to routinely check their health status and potentially discover asymptomatic conditions is simply a time management decision. [Table 2, Respondents in good health, Time preference: Marginal effect = 0.02763, SE = 0.01380, significant at 5% level].  

### Reproduction data source(s)

https://www.rand.org/well-being/social-and-behavioral-policy/data/FLS/IFLS/ifls4.html  

### Description of reproduction data 

The fourth wave of the Indonesia Family Life Survey (IFLS4) is a continuation of IFLS, expanding the panel to 2007/2008. Downloaded from the link in the cell above on 2022-01-06; uploaded to OSF to ‘Data/data-raw’ directory.      

Selecting respondents in good health only.  

Dataset consists of `r number(nrow(data_good_health), big.mark = ",")` observations.  

### Model 

See model `claim_21v1on` above.    

```{r}
summary(margins(claim_21v1on))
```

### Primary reproduction criteria

```{r}
primary <- read_sheet(sheet_location,
                      range = "Claim_ID_dx1x5w!1:6") %>% 
  slice(-1) %>% 
  mutate(`Reproduction result` = "")

primary$`Reproduction result`[1] <- number(nrow(data_good_health))

primary$`Reproduction result`[3] <- summary(margins(claim_21v1on)) %>% 
  filter(factor == "time") %>% 
  select(AME) %>% 
  mutate(AME = number(AME, accuracy = 0.000001)) %>% 
  pluck(1)

primary %>% 
  kableExtra::kable()
```

### Secondary reproduction criteria

```{r}
secondary <- read_sheet(sheet_location,
                        range = "Claim_ID_dx1x5w!1:10") %>% 
  slice(8:9) %>% 
  mutate(`Reproduction result` = "")

secondary$`Reproduction result`[1] <- summary(margins(claim_21v1on)) %>% 
  filter(factor == "time") %>% 
  select(SE) %>% 
  mutate(SE = number(SE, accuracy = 0.000001)) %>% 
  pluck(1)

secondary$`Reproduction result`[2] <- summary(margins(claim_21v1on)) %>% 
  filter(factor == "time") %>% 
  select(p) %>% 
  mutate(p = number(p, accuracy = 0.000001)) %>% 
  pluck(1)

secondary %>% 
  kableExtra::kable()
```

### Analyst success criteria

All outcomes meet the precise reproduction criteria

### Reproduction outcome

Based on the criteria above, the claim **did not** reproduce.  


## Claim ID: 8rlrky

### Coded claim 4 text (original paper)

> As predicted, we do observe different effects, both in magnitude and significance for generally healthy and generally unhealthy respondents. The level of education matters for all respondents, but with different magnitudes. More specifically, for individuals in poorer general health, education has a larger impact than for individuals in good health who possess no obvious impairing symptoms. [Table 4, Respondents in good health, Years of Education: Marginal effect = 0.00625, SE = 0.00281, significant at 5% level; Respondents in poor health, Years of Education: Marginal effect = 0.01407, SE = 0.00499, significant at 1% level]    

### Reproduction data source(s)

https://www.rand.org/well-being/social-and-behavioral-policy/data/FLS/IFLS/ifls4.html  

### Description of reproduction data 

The fourth wave of the Indonesia Family Life Survey (IFLS4) is a continuation of IFLS, expanding the panel to 2007/2008. Downloaded from the link in the cell above on 2022-01-06; uploaded to OSF to ‘Data/data-raw’ directory.  

Selecting diagnosed respondents in good health only.  

```{r}
data_diag <- data %>% 
  filter(underdiagnosed == 0) %>% 
  select(-underdiagnosed) %>% 
  filter(poor_health == 0) %>% 
  select(-poor_health)
```

Dataset consists of `r number(nrow(data_diag), big.mark = ",")` observations.  

### Model 

```{r}
claim_8rlrky <- glm(bp_month2 ~ 
                      age + age2 + education +
                      expenditures_log + time + risk +
                      travel_time + sex, 
                    family = binomial(link = "probit"), 
                    data = data_diag)

# summary(claim_8rlrky)
summary(margins(claim_8rlrky))
```

### Primary reproduction criteria

```{r}
primary <- read_sheet(sheet_location,
                      range = "Claim_ID_8rlrky!1:6") %>% 
  slice(-1) %>% 
  mutate(`Reproduction result` = "")

primary$`Reproduction result`[1] <- number(nrow(data_diag))

primary$`Reproduction result`[3] <- summary(margins(claim_8rlrky)) %>% 
  filter(factor == "education") %>% 
  select(AME) %>% 
  mutate(AME = number(AME, accuracy = 0.000001)) %>% 
  pluck(1)

primary %>% 
  kableExtra::kable()
```

### Secondary reproduction criteria

```{r}
secondary <- read_sheet(sheet_location,
                        range = "Claim_ID_8rlrky!1:10") %>% 
  slice(8:9) %>% 
  mutate(`Reproduction result` = "")

secondary$`Reproduction result`[1] <- summary(margins(claim_8rlrky)) %>% 
  filter(factor == "education") %>% 
  select(SE) %>% 
  mutate(SE = number(SE, accuracy = 0.000001)) %>% 
  pluck(1)

secondary$`Reproduction result`[2] <- summary(margins(claim_8rlrky)) %>% 
  filter(factor == "education") %>% 
  select(p) %>% 
  mutate(p = number(p, accuracy = 0.000001)) %>% 
  pluck(1)

secondary %>% 
  kableExtra::kable()
```

### Analyst success criteria

All outcomes meet the precise reproduction criteria

### Reproduction outcome

Based on the criteria above, the claim **did not** reproduce.  


<!-- ----------------------------------------------------- -->

# Final remarks

All in all I would consider this paper to be pretty much impossible to be faithfully reproduced. To start with - the paper does not even report which software was used in the analyses. More worryingly, the amount of undocumented and undescribed steps taken in the data preparation itself could lead to virtually infinite amount of slightly different analyses that could be conducted.  

The preparation of 'simple' years of education variable - definitions, criteria, handling of missings is already extremely difficult with lack of code and laconic description provided in the paper. More elaborate variables such as time preference construct for instance present possibilities to again create multiple versions of the same dataset - which risk was taken into account - A or B? maybe both? what happened to missing information?  

I cannot question validity of the paper itself. But I think it's safe to assume that without any additional information from the authors of the study - either in a descriptive from properly defining the steps taken or in the form of computer code, this study is impossible to be reproduced.  

<!-- ----------------------------------------------------- -->

# List of files

1. Raw data in Stata's dta format are stored in `data-raw` folder (preserving original names; available from RAND the data provider [website](https://www.rand.org/well-being/social-and-behavioral-policy/data/FLS/IFLS/ifls4.html); requires registration; stored on OSF project website). Files used to construct final dataset:  

- bus1_1.dta  
- bus1_2.dta  
- b3b_cd3.dta  
- b1_ks2.dta  
- b3b_kk1.dta  
- b1_pp.dta  
- b3a_dl1.dta  
- b3a_si.dta  
- b3b_rj3.dta  
- b1_cov.dta  

2. Raw data are processed and combined into a final dataset using R markdwon script `02_data-management.Rmd` stored in `analyses` folder of the project (with output of the script stored in `02_data-management.html` file in `docs` folder).  

3. Final dataset `data.Rds` is stored in `data` folder.  

4. Analyses are conducted and documented in `03_data-analysis.Rmd` script (with output of the script stored in `03_data-analysis.html` file in `docs` folder).  

<!-- ----------------------------------------------------- -->

# References

Kim Y, Radoias V (2016) Education, individual time preferences, and asymptomatic disease detection. *Social Science & Medicine* **150**, 15-22, https://doi.org/10.1016/j.socscimed.2015.11.051.  

Ng J (2012) Elicited Risk and Time Preferences: The Role of Demographics, Cognition, and Interviewers. Working paper. University of Southern California. http://static1.1.sqspcdn.com/static/f/884336/20238159/1347396545633/ng_dev_seminar9-10.pdf  

<!-- ----------------------------------------------------- -->

# Computing Environment

```{r echo=FALSE, results='asis'}
report(sessionInfo())
```
